Moved Enum classes to new package util.enums
